ux improvements
