+++ website/src/components/themeButton/ThemeButton.tsx
@@ -6,11 +6,7 @@ const ThemeButton: FunctionComponent = () => {
   const [isDarkMode, setIsDarkMode] = useState(false);
 
   useEffect(() => {
-    if (
-      localStorage.theme === "dark" ||
-      (!("theme" in localStorage) &&
-        window.matchMedia("(prefers-color-scheme: dark)").matches)
-    ) {
+    if (localStorage.theme === "dark" || !("theme" in localStorage)) {
       document.documentElement.classList.add("dark");
       setIsDarkMode(true);
     } else {
